Add delete order with brains as well as get orders by id and get all orders

+++ src/api/models/orderModel.js
@@ -12,12 +12,18 @@ const addOrder = async (body, userId) => {
   return { message: "Order places" };
 };
 
-const getAllOrders = async () => {
+const getAllOrders = async (user) => {
+  if (user.access !== "admin") {
+    return false;
+  }
   const [orders] = await promisePool.query("SELECT * from orders");
   return orders;
 };
 
-const getUserOrder = async (id) => {
+const getUserOrder = async (id, user) => {
+  if (user.access !== "admin" || user.access !== "user") {
+    return false;
+  }
   const [rows] = await promisePool.execute(
     "SELECT * from orders WHERE orderer = ?",
     [id],
@@ -28,19 +34,18 @@ const getUserOrder = async (id) => {
   return rows;
 };
 
-const delOrder = async (id) => {
+const delOrder = async (id, user) => {
+  if (user.access !== "admin") {
+    return { message: "Unauthorized" };
+  }
   const sql = "DELETE FROM orders WHERE id = ?";
   const [result] = await promisePool.execute(sql, [id]);
-  if (result.affectedRows > 0) {
-    return true;
-  } else {
-    return false;
-  }
+  return result.affectedRows !== 0;
 };
 
 const deliverOrder = async (id) => {
   try {
-    const sql = 'UPDATE orders SET status = ? WHERE id = ?';
+    const sql = "UPDATE orders SET status = ? WHERE id = ?";
     const status = 1;
     const [result] = await promisePool.execute(sql, [status, id]);
     if (result.affectedRows > 0) {
@@ -49,7 +54,7 @@ const deliverOrder = async (id) => {
       return false;
     }
   } catch (error) {
-    console.error('Error delivering order:', error);
+    console.error("Error delivering order:", error);
     return false; // Return false to indicate that there was an error
   }
 };
